# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedView": {
        "properties": {
          "auto_download": {
            "title": "Auto Download",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Actor"
              },
              {
                "type": "null"
              }
            ]
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemView"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null while the feed is healthy."
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_check_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null while the feed is disabled.",
            "title": "Next Check At"
          },
          "refresh_minutes": {
            "title": "Refresh Minutes",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The feed's own title, from its last fetch.",
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Actor"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "description": "The address with passkeys and tokens masked. Admins can read it whole at GET /feeds/{id}/url.",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "enabled",
          "refresh_minutes",
          "auto_download",
          "title",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by",
          "last_checked_at",
          "next_check_at",
          "last_error"
        ],
        "title": "FeedView",
        "type": "object"
      },
...
      "MatchFilter": {
        "enum": [
          "all",
          "matched",
          "unmatched"
        ],
        "title": "MatchFilter",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Charon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/feeds": {
      "get": {
        "operationId": "listFeeds",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedView"
                  },
                  "title": "Response Listfeeds",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Feeds",
        "tags": [
          "feeds"
        ]
      },
      "post": {
        "description": "Subscribe, and fetch the feed once right away.\n\nA feed that can't be fetched is still added, with `last_error` saying why.",
        "operationId": "createFeed",
        "parameters": [
          {
            "description": "Any unique string, e.g. a UUID. Repeating a request with the same key returns what the first one created (with status 200) instead of creating it again. Keys belong to the API key that sent them and are remembered for a day.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any unique string, e.g. a UUID. Repeating a request with the same key returns what the first one created (with status 200) instead of creating it again. Keys belong to the API key that sent them and are remembered for a day.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedView"
                }
              }
            },
            "description": "Repeated request"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflicts with the resource's current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Feed",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/items": {
      "get": {
        "description": "Items from every feed, newest published first, each checked against the current rules.",
        "operationId": "listFeedItems",
        "parameters": [
          {
            "in": "query",
            "name": "feed_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Id"
            }
          },
          {
            "in": "query",
            "name": "match",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MatchFilter",
              "default": "all"
            }
          },
          {
            "description": "Only items nobody has seen yet.",
            "in": "query",
            "name": "unseen",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only items nobody has seen yet.",
              "title": "Unseen",
              "type": "boolean"
            }
          },
          {
            "description": "Text the name contains.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text the name contains.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Feed Items",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/items/seen": {
      "post": {
        "description": "Mark exactly these items seen, e.g. the ones a list showed. Unknown ones are ignored.",
        "operationId": "markFeedItemsSeen",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkSeenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkSeenView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Mark Feed Items Seen",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/items/seen-all": {
      "post": {
        "description": "Mark every unseen item in a view seen (\"Mark all seen\"), up to `up_to`.\n\nThe view's `feed_id`, `match` and `q` apply, so items it hides stay new.",
        "operationId": "markAllFeedItemsSeen",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAllSeenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkSeenView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Mark All Feed Items Seen",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/items/{info_hash}": {
      "get": {
        "operationId": "getFeedItem",
        "parameters": [
          {
            "in": "path",
            "name": "info_hash",
            "required": true,
            "schema": {
              "title": "Info Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Feed Item",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/items/{info_hash}/download": {
      "post": {
        "description": "Download an item. Answers 200 with the existing job if the torrent is already in Charon.",
        "operationId": "downloadFeedItem",
        "parameters": [
          {
            "in": "path",
            "name": "info_hash",
            "required": true,
            "schema": {
              "title": "Info Hash",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Already in Charon"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The download backend failed or is unreachable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Download Feed Item",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/preview": {
      "post": {
        "description": "Fetch and read a feed without subscribing: its title, counts and newest items.",
        "operationId": "previewFeed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPreviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Preview Feed",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/refresh": {
      "post": {
        "description": "Fetch every enabled feed now; paused ones are skipped. Problems go in `last_error`.",
        "operationId": "refreshAllFeeds",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedView"
                  },
                  "title": "Response Refreshallfeeds",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Refresh All Feeds",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/summary": {
      "get": {
        "description": "How many items nobody has seen yet, in total and per feed.",
        "operationId": "summarizeFeeds",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedSummaryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Summarize Feeds",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/{feed_id}": {
      "delete": {
        "description": "Unsubscribe. Items only this feed listed are forgotten; downloads are untouched.",
        "operationId": "deleteFeed",
        "parameters": [
          {
            "in": "path",
            "name": "feed_id",
            "required": true,
            "schema": {
              "title": "Feed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Feed",
        "tags": [
          "feeds"
        ]
      },
      "get": {
        "operationId": "getFeed",
        "parameters": [
          {
            "in": "path",
            "name": "feed_id",
            "required": true,
            "schema": {
              "title": "Feed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Feed",
        "tags": [
          "feeds"
        ]
      },
      "put": {
        "description": "Change a feed's settings. Changing its address (`url`) takes an admin key.",
        "operationId": "updateFeed",
        "parameters": [
          {
            "in": "path",
            "name": "feed_id",
            "required": true,
            "schema": {
              "title": "Feed Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The key's role does not allow this action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Feed",
        "tags": [
          "feeds"
        ]
      }
    },
...
  }
}